+++ test/unit/index.spec.ts
@@ -50,3 +50,9 @@ import './firestore/firestore.spec';
 // InstanceId
 import './instance-id/instance-id.spec';
 import './instance-id/instance-id-request.spec';
+
+// ProjectManagement
+import './project-management/project-management.spec';
+import './project-management/project-management-api-request.spec';
+import './project-management/android-app.spec';
+import './project-management/ios-app.spec';