+++ CHANGELOG.md
@@ -18,6 +18,9 @@
 - [changed] Upgraded Cloud Firestore client to v0.18.0.
 - [added] Exposed the `CollectionReference`, `WriteBatch`, `WriteResult` and
   `QueryDocumentSnapshot` types from the `admin.firestore` namespace.
+- [added] A new `ProjectManagement` service, which includes the ability to
+  create, list, and get details about Android and iOS apps associated with your
+  Firebase Project.
 
 # v6.0.0